iOS picker fixes. Excess code deleted.
