+++ Assets/Puppet/FilePicker/AndroidFilePicker.cs
@@ -1,4 +1,7 @@
-﻿using Assets.Puppet.Plugins.Android.FilePickerManager;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+using Assets.Puppet.Plugins.Android.FilePickerManager;
 
 public class AndroidFilePicker : IFilePicker
 {
