--- Assets/Puppet/FilePicker/AttachmentsController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e2bf9bb1b3998a4bb9c9e3452e7da3f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
