+++ Assets/Puppet/FilePicker/IFilePicker.cs
@@ -1,4 +1,7 @@
-﻿internal interface IFilePicker
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+internal interface IFilePicker
 {
     void Show();
 }
