+++ Assets/Puppet/FilePicker/IOSFilePicker.cs
@@ -1,4 +1,7 @@
-﻿using System.Runtime.InteropServices;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+using System.Runtime.InteropServices;
 
 internal class IOSFilePicker : IFilePicker
 {
