+++ Assets/Puppet/Plugin/iOS/FilePickerManager/FilePicker.h
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 #import <UIKit/UIKit.h>
 
 @interface FilePicker : NSObject<UIDocumentPickerDelegate>
